# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ZVideo API
  version: 1.0.0
  description: REST API for a video hosting platform.
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Delete current user account
      operationId: deleteCurrentUser
      responses:
        '204':
          description: Account deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me/notification-settings:
    patch:
      tags: [Users]
      summary: Update current user's notification settings
      operationId: updateCurrentUserNotificationSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSettingsRequest'
      responses:
        '204':
          description: Notification settings updated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users:
    get:
      tags: [Users]
      summary: List users
      description: Requires admin role.
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Users page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{userId}/ban:
    post:
      tags: [Users]
      summary: Ban a user
      description: Requires admin role.
      operationId: banUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User banned.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{userId}/unban:
    post:
      tags: [Users]
      summary: Unban a user
      description: Requires admin role.
      operationId: unbanUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User unbanned.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{userId}/role:
    put:
      tags: [Users]
      summary: Replace user role
      description: Requires admin role.
      operationId: replaceUserRole
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeUserRoleRequest'
      responses:
        '204':
          description: Role changed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{userId}/channel:
    get:
      tags: [Channels]
      security: []
      summary: Get a user's channel
      operationId: getUserChannel
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User's channel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /channels/{channelId}/playlists:
    get:
      tags: [Playlists]
      security: []
      summary: List channel playlists
      operationId: listChannelPlaylists
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Playlists page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistPage'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Playlists]
      summary: Create playlist in channel
      operationId: createPlaylist
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaylistCreateRequest'
      responses:
        '201':
          description: Playlist created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/me/playlists:
    get:
      tags: [Playlists]
      summary: List current user's playlists
      operationId: listMyPlaylists
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Playlists page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /playlists/{playlistId}:
    get:
      tags: [Playlists]
      security: []
      summary: Get playlist
      operationId: getPlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
      responses:
        '200':
          description: Playlist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Playlists]
      summary: Update playlist
      operationId: updatePlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaylistUpdateRequest'
      responses:
        '200':
          description: Playlist updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Playlists]
      summary: Delete playlist
      operationId: deletePlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
      responses:
        '204':
          description: Playlist deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /playlists/{playlistId}/videos:
    get:
      tags: [Playlists]
      security: []
      summary: List playlist videos
      operationId: listPlaylistVideos
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Playlist items page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistItemPage'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Playlists]
      summary: Add video to playlist
      operationId: addVideoToPlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPlaylistVideoRequest'
      responses:
        '204':
          description: Video added.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /playlists/{playlistId}/videos/{videoId}:
    patch:
      tags: [Playlists]
      summary: Update video position in playlist
      operationId: updatePlaylistVideoPosition
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
        - $ref: '#/components/parameters/VideoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlaylistVideoPositionRequest'
      responses:
        '204':
          description: Position updated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Playlists]
      summary: Remove video from playlist
      operationId: removeVideoFromPlaylist
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
        - $ref: '#/components/parameters/VideoId'
      responses:
        '204':
          description: Video removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    ChannelId:
      name: channelId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    VideoId:
      name: videoId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
    PlaylistId:
      name: playlistId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Request validation or domain rule failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidUsername:
              summary: Invalid username
              value:
                error:
                  code: INVALID_USERNAME
                  message: The username is invalid
            invalidUserEmail:
              summary: Invalid email
              value:
                error:
                  code: INVALID_USER_EMAIL
                  message: The email is invalid
            weakPassword:
              summary: Weak password
              value:
                error:
                  code: WEAK_PASSWORD
                  message: The password is too weak
            playlistNameEmpty:
              summary: Empty playlist name
              value:
                error:
                  code: PLAYLIST_NAME_EMPTY
                  message: Playlist name cannot be empty
            playlistVideoChannelMismatch:
              summary: Playlist and video belong to different channels
              value:
                error:
                  code: PLAYLIST_VIDEO_CHANNEL_MISMATCH
                  message: Video belongs to another channel
            selfSubscription:
              summary: User cannot subscribe to own channel
              value:
                error:
                  code: SELF_SUBSCRIPTION
                  message: Cannot subscribe to your own channel
            invalidNotificationSettings:
              summary: Invalid notification settings
              value:
                error:
                  code: INVALID_NOTIFICATION_SETTINGS
                  message: Notification settings payload is invalid
            roleNotFound:
              summary: Role not found
              value:
                error:
                  code: ROLE_NOT_FOUND
                  message: Role not found
            communityPostContentEmpty:
              summary: Empty community post
              value:
                error:
                  code: COMMUNITY_POST_CONTENT_EMPTY
                  message: Community post content cannot be empty
            communityCommentContentEmpty:
              summary: Empty community comment
              value:
                error:
                  code: COMMUNITY_COMMENT_CONTENT_EMPTY
                  message: Community comment content cannot be empty
    Unauthorized:
      description: Authentication failed or token is missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidCredentials:
              summary: Invalid login credentials
              value:
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password
            invalidToken:
              summary: Invalid or expired token
              value:
                error:
                  code: INVALID_TOKEN
                  message: Invalid or expired token
            missingAuthorization:
              summary: Missing bearer token
              value:
                error:
                  code: UNAUTHORIZED
                  message: user not authorized
    Forbidden:
      description: Authenticated user is not allowed to perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              summary: Insufficient permissions
              value:
                error:
                  code: FORBIDDEN
                  message: You don't have permission to access this resource
            userBanned:
              summary: User account is banned
              value:
                error:
                  code: USER_BANNED
                  message: User account is banned
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            userNotFound:
              summary: User not found
              value:
                error:
                  code: USER_NOT_FOUND
                  message: User not found
            channelNotFound:
              summary: Channel not found
              value:
                error:
                  code: CHANNEL_NOT_FOUND
                  message: Channel not found
            videoNotFound:
              summary: Video not found
              value:
                error:
                  code: VIDEO_NOT_FOUND
                  message: Video not found
            playlistNotFound:
              summary: Playlist not found
              value:
                error:
                  code: PLAYLIST_NOT_FOUND
                  message: Playlist not found
            ratingNotFound:
              summary: Rating not found
              value:
                error:
                  code: RATING_NOT_FOUND
                  message: Rating not found
            commentNotFound:
              summary: Comment not found
              value:
                error:
                  code: COMMENT_NOT_FOUND
                  message: Comment not found
            communityPostNotFound:
              summary: Community post not found
              value:
                error:
                  code: COMMUNITY_POST_NOT_FOUND
                  message: Community post not found
            communityCommentNotFound:
              summary: Community comment not found
              value:
                error:
                  code: COMMUNITY_COMMENT_NOT_FOUND
                  message: Community comment not found
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: ERROR_CODE
        message:
          type: string
          example: Human-readable error message
...
    Role:
      type: object
      required: [id, name, isDefault]
      properties:
        id:
          type: integer
        name:
          $ref: '#/components/schemas/RoleName'
        isDefault:
          type: boolean
    RoleName:
      type: string
      enum: [admin, moderator, user]
    User:
      type: object
      required: [id, username, email, isActive, notificationsEnabled, createdAt, updatedAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        isActive:
          type: boolean
        notificationsEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        role:
          $ref: '#/components/schemas/Role'
    NotificationSettingsRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    ChangeUserRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: '#/components/schemas/RoleName'
    Channel:
      type: object
      required: [id, userId, ownerUsername, name, description, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        ownerUsername:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
...
    VideoStatus:
      type: string
      enum: [pending, ready]
...
    Playlist:
      type: object
      required: [id, channelId, name, description, createdAt]
      properties:
        id:
          type: integer
        channelId:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    PlaylistCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        description:
          type: string
          default: ''
    PlaylistUpdateRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        description:
          type: string
    AddPlaylistVideoRequest:
      type: object
      required: [videoId]
      properties:
        videoId:
          type: integer
          minimum: 1
    UpdatePlaylistVideoPositionRequest:
      type: object
      required: [position]
      properties:
        position:
          type: integer
          minimum: 1
    PlaylistItem:
      type: object
      required: [playlistId, videoId, number, addedAt, videoTitle, channelName, videoStatus]
      properties:
        playlistId:
          type: integer
        videoId:
          type: integer
        number:
          type: integer
          minimum: 1
        addedAt:
          type: string
          format: date-time
        videoTitle:
          type: string
        channelName:
          type: string
        videoStatus:
          $ref: '#/components/schemas/VideoStatus'
...
    PageMeta:
      type: object
      required: [limit, offset, totalCount]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        totalCount:
          type: integer
          format: int64
          minimum: 0
    UserPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/User'
...
    PlaylistPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Playlist'
    PlaylistItemPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/PlaylistItem'
...